Improve UpdateTotals API to calculate all totals automatically
Previously, UpdateTotals() required users to manually call another method
first to calculate allowance/charge totals, which was confusing and error-prone.
Now UpdateTotals() truly updates ALL totals as users would expect.

CHANGES:
- UpdateTotals() now automatically calculates AllowanceTotal & ChargeTotal
- Implemented as internal helper: updateAllowancesAndCharges() (private)
- Reordered calculation steps to match logical flow:
  1. LineTotal (from invoice lines)
  2. AllowanceTotal & ChargeTotal (from document-level items)
  3. TaxTotal (from VAT breakdown)
  4. TaxBasisTotal (LineTotal - Allowances + Charges)
  5. GrandTotal (TaxBasisTotal + TaxTotal)
  6. DuePayableAmount (GrandTotal - Prepaid + Rounding)

- Reorganized comments to show business rules in calculation order
- Updated existing tests to provide actual allowance/charge items
- Added comprehensive integration tests

BENEFITS:
- Simpler API: Users only call UpdateTotals() once
- More intuitive: Function name matches actual behavior
- Less error-prone: Can't forget to calculate allowances/charges
- Better documented: Comments follow calculation flow
